Refactor, Document, and Improve Robustness of mymc-bot
This commit introduces several improvements to the mymc-bot project:

- **Refactoring**: The code has been restructured into `bot` and `plugins` modules for better organization and readability.
- **Documentation**: Comprehensive `rustdoc`-style documentation has been added in Indonesian to all modules, structs, and functions.
- **Robustness**:
    - State management has been improved by removing a redundant `RefCell` and simplifying the `TaskModel`.
    - The hardcoded owner UUID is now a constant.
    - The previously empty `tick_commands` function has a basic implementation.
    - `BotPlugins` for auto-reconnect and auto-respawn have been enabled.
- **Build**: A `rust-toolchain.toml` file has been added to ensure the project is built with the required nightly toolchain.
