+++ Projects/mymc-bot/src/bot/mod.rs
@@ -1,0 +1,4 @@
+//! Modul ini berisi logika inti dari bot.
+
+pub mod component;
+pub mod handler;
