+++ Projects/mymc-bot/src/main.rs
@@ -1,30 +1,21 @@
+//! Bot Minecraft sederhana yang dibuat dengan Azalea.
+
 use azalea::{app::{PluginGroup, PluginGroupBuilder}, auto_reconnect::AutoReconnectPlugin, auto_respawn::AutoRespawnPlugin, prelude::*};
 
-// Modules
-mod component;
-mod handler;
+mod bot;
+mod plugins;
 
-// re-export
-use component::BotComponent;
-use handler::handle;
+use bot::component::BotComponent;
+use bot::handler::handle;
+use plugins::BotPlugins;
 
-#[allow(dead_code)]
-struct BotPlugins;
-
-impl PluginGroup for BotPlugins {
-    fn build(self) -> PluginGroupBuilder {
-        PluginGroupBuilder::start::<Self>()
-        .add(AutoReconnectPlugin)
-        .add(AutoRespawnPlugin)
-    }
-}
-
+/// Fungsi utama untuk menjalankan bot.
 #[tokio::main]
 async fn main() {
     let account = Account::offline("itzyuurzbot");
     ClientBuilder::new()
     .set_handler(handle)
-    // .add_plugins(BotPlugins)
+    .add_plugins(BotPlugins)
     .set_state(BotComponent::default())
     .start(account, "itzyuurz.aternos.me:11068")
     .await.unwrap();
